Conflicts:
	.gitignore
	lawa/methods/adaptive/algorithms/security_zone.h
	lawa/methods/adaptive/datastructures/index.h
	lawa/methods/adaptive/datastructures/matrixoperations.h
	lawa/methods/adaptive/datastructures/matrixoperations.tcc
	lawa/methods/adaptive/operators/adaptivespacetimepdeoperator1d.h
	lawa/methods/adaptive/operators/adaptivespacetimepdeoperator1d.tcc
	lawa/methods/adaptive/operators/operators.h
	lawa/methods/adaptive/operators/pdeoperators2d/adaptivehelmholtzoperator2d.tcc
	lawa/methods/adaptive/operators/spacetimeoperators/adaptivespacetimepdeoperator1d.h
	lawa/methods/adaptive/operators/spacetimeoperators/adaptivespacetimepdeoperator1d.tcc
	lawa/methods/adaptive/operators/spacetimeoperators/adaptivespacetimeweightedconvectionoperator1d.h
	lawa/methods/adaptive/operators/spacetimeoperators/adaptivespacetimeweightedconvectionoperator1d.tcc
	lawa/methods/adaptive/operators/spacetimeoperators/spacetimeoperators.h

+++ lawa/operators/deltas.tcc
@@ -2,7 +2,7 @@ namespace lawa {
 
 template <typename T, typename Basis>
 GeMatrix<FullStorage<T,ColMajor> >
-computeDeltas(const Basis &basis, int j, int k, XType e)
+computeDeltas(const Basis &basis, int j, long k, XType e)
 {
     GeMatrix<FullStorage<T,ColMajor> > ret;
     const typename Basis::BasisFunctionType &varphi=basis.generator(e); //either B-Spline or Wavelet
