+++ lawa/operators/pdeoperators1d/convectionoperator1d_pg.tcc
@@ -28,8 +28,8 @@ ConvectionOperator1D_PG<T, TrialBasis, TestBasis>::ConvectionOperator1D_PG(const
 
 template <typename T, typename TrialBasis, typename TestBasis>
 T
-ConvectionOperator1D_PG<T, TrialBasis, TestBasis>::operator()(XType xtype1, int j1, int k1,
-                                                           XType xtype2, int j2, int k2) const
+ConvectionOperator1D_PG<T, TrialBasis, TestBasis>::operator()(XType xtype1, int j1, long k1,
+                                                              XType xtype2, int j2, long k2) const
 {   
     // v * u_x
     return integral(j1, k1, xtype1, 0, j2, k2, xtype2, 1);
