+++ lawa/righthandsides/separablerhs3d.tcc
@@ -19,9 +19,9 @@ SeparableRHS3D<T, Basis3D>::SeparableRHS3D
 
 template<typename T, typename Basis3D>
 T
-SeparableRHS3D<T, Basis3D>::operator()(XType xtype_x, int j_x, int k_x,
-                                       XType xtype_y, int j_y, int k_y,
-                                       XType xtype_z, int j_z, int k_z) const
+SeparableRHS3D<T, Basis3D>::operator()(XType xtype_x, int j_x, long k_x,
+                                       XType xtype_y, int j_y, long k_y,
+                                       XType xtype_z, int j_z, long k_z) const
 {
     T val_x = 0;
     T val_y = 0;
